Blueprints internals – split the `common.ts` file into smaller files
This small commit brings a part of #851 into trunk for easier review.

 ## Testing instructions

Confirm the CI tests pass
