+++ packages/playground/blueprints/src/index.ts
@@ -32,6 +32,8 @@ export type {
 	VFSResource,
 } from './lib/resources';
 
+export { wpContentFilesExcludedFromExport } from './lib/utils/wp-content-files-excluded-from-exports';
+
 /**
  * @deprecated This function is a no-op. Playground no longer uses a proxy to download plugins and themes.
  *             To be removed in v0.3.0
