+++ packages/playground/blueprints/src/lib/steps/import-wp-content.ts
@@ -1,8 +1,8 @@
 import { StepHandler } from '.';
 import { unzip } from './unzip';
 import { dirname, joinPaths, phpVar } from '@php-wasm/util';
-import { wpContentFilesExcludedFromExport } from './common';
 import { UniversalPHP } from '@php-wasm/universal';
+import { wpContentFilesExcludedFromExport } from '../utils/wp-content-files-excluded-from-exports';
 
 /**
  * @inheritDoc importWordPressFiles
